+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Behavior of File Modifier to return original data when manipulation fails silently
 - Generic fieldtype passes `content_id` along to core fieldtype handler and triggers `pre_process` hook.
+- Error checking and handling for GraphQL compatible fieldtypes
 
 ## [1.1.2] - 2023-07-27