+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Generic fieldtype passes `content_id` along to core fieldtype handler and triggers `pre_process` hook.
 - Error checking and handling for GraphQL compatible fieldtypes
 - Trigger `core_boot` hook during GraphQL requests
+- Handling of blank headers in HTTP responses
 
 ## [1.1.2] - 2023-07-27