+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Artisan command to proxy ExpressionEngine's `eecli` commands in the context of Laravel and Coilpack after they have fully booted.  This command is available through `php artisan eecli`.
+- Parameter to Channel Entries Tag for eager loading relationships.  This allows for easier inclusion of custom field data with `exp.channel.entries({with: 'data'})`
 ### Fixed
 
 - Behavior of File Modifier to return original data when manipulation fails silently
@@ -12,6 +13,7 @@
 - Error checking and handling for GraphQL compatible fieldtypes
 - Trigger `core_boot` hook during GraphQL requests
 - Handling of blank headers in HTTP responses
+- Custom field orderby clauses when it uses legacy field data storage
 
 ## [1.1.2] - 2023-07-27