+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Added
 
+- The ability for ExpressionEngine to utilize Laravel cache adapters
 - Artisan command to proxy ExpressionEngine's `eecli` commands in the context of Laravel and Coilpack after they have fully booted.  This command is available through `php artisan eecli`.
 - Parameter to Channel Entries Tag for eager loading relationships.  This allows for easier inclusion of custom field data with `exp.channel.entries({with: 'data'})`
+
 ### Fixed
 
 - Behavior of File Modifier to return original data when manipulation fails silently